\name{mahDis}
\alias{mahDis}
\alias{mahalanobisGROUPS}
\alias{mahalanobisgroups}
\alias{mahdis}
\alias{mahDIS}
\title{Mahalanobis Distance Between Two Groups}
\description{
Calculates Mahalanobis distance between the groups centroids median position.
}
\usage{
mahDis(data, group)
}
\arguments{
  \item{data}{is a matrix containing point positions.  The matrix has two columns. Each row of the matrix gives one point.}

  \item{group}{is a vector specifying the group each point belongs to. Notice that this method calculates distances of two groups by a time.}
}
